mgs: enable AppKM in MGS if policy is enabled
This is the first CL to enable AppKM in MGS. AppKM are reported for MGS
only if `UrlKeyedAnonymizedDataCollectionEnabled` policy is enabled.

`UrlKeyedAnonymizedDataCollectionEnabled` policy is not available in
the admin console, so current changes are hidden behind this policy.

With this change AppKM is reported only for long sessions (>2 hours).
Metrics for shorter sessions are erased on user logout, since MGS
clean up all user preferences. That will be fixed in the following
CLs.

Also AppKM will be restricted to admins-configured apps only in the
following CLs.

Bug: b/303610964
Change-Id: If714021abe5502b2fd0f0ddb0e09181bc2905ce5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4928548
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Edman Anjos <edman@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Irina Fedorova <irfedorova@google.com>
Cr-Commit-Position: refs/heads/main@{#1212381}