[SHv2] Add SafetyHubService result for password checks
This CL adds PasswordStatusCheckResult class to be used when showing
notification for password leaks in three dot menu. Since
PasswordStatusCheckService is not a SafetyHubService, some functions
like GetCachedResult and GetResultFromDictValue are implemented
additionally to mimic behavior of SafetyHubService.

A follow up CL will add the functionality to show passwords in the
three dot menu.

Change-Id: I0d3dc1f419136d4d23651ca34d238aa99c80169b
Bug: 1443466
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4898466
Commit-Queue: Side YILMAZ <sideyilmaz@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212387}