[omnibox][webui] Size by ResizeObserver not EnableSizingFromWebContents
This CL eliminates use of EnableSizingFromWebContents and the related
FrameSizeChanged method because the sizing behavior observed on
different systems was unpredictable and generally incorrect. With this
change, the browser's ResizeObserver API is used to detect the actual
size of the webui dropdown content element. The actual size is plumbed
through the RealboxHandler to the OmniboxPopupPresenter so it can size
the containing widget accurately.

Bug: 1479181
Change-Id: I3caa1dbbba7d3910db2718573db2a702e7bf1334
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4923182
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: John Lee <johntlee@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212402}