[Autofill] Pass AutofillDriver to ExtractForm() callback
This CL parameterizes the ExtractForm() callback in the browser process
with the AutofillDriver that manages the form.

Renderer-initiated events get routed to the AutofillDriver that manages
the flattened form.

With ExtractForm(), the flattened form reaches whichever driver called
ExtractForm(). This is not necessarily the driver that manages the
flattened form.

With this CL, the handler now gets access to the managing driver. Also,
this parameter is intended to serve as a warning to callers of
ExtractForm() to be careful about the drivers. In particular, they
should not add the received form to the wrong driver's
AutofillManager's cache.

Bug: 1490670, 1484810
Change-Id: I296422967d8f10f70562232c1d5740a638dd1ee6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4952976
Reviewed-by: Dominic Battre <battre@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1212407}