Fix renderer content settings (part 4 / N)
This CL has no intended functional change.

In this CL, the browser computes and sends content settings for each
frame navigation. Note that the data is not yet consumed by the
renderer, and the browser continues to send the raw rules as well.

Change-Id: I89929091c792b4542e448d0aaaedfe41bd8a77ac
Bug: 1208547
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4933050
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212425}