IndexedDB: lazily recalculate usage for quota
Quota clients such as IndexedDB typically provide the quota manager
with an initial usage figure and later update it as needed via
`NotifyBucketModified` and a provided size delta, measured in bytes.
For IndexedDB, predicting changes in usage just by looking at
in-flight transactions is very hard, so after every operation that
potentially changes the usage, it re-queries the disk to
get the total size of the folders where it stores data (LevelDB files
and blobs). Somewhat backwardly, it then calculates a delta by
subtracting the old value from the new value, and passes that delta to
the quota manager. This is a very inefficient system because it reads
from disk quite often (e.g. after each one of 1000 sequential
transactions), and in so doing blocks queued IDB actions and
transactions.

It's apparent that the design of the quota manager-client
relationship was not conceived with IndexedDB in mind. This change
modifies the contract between client and manager such that clients
*may* opt to simply dirty the usage cache that's kept by the manager.
Instead of passing a delta in bytes, the client (in this case IDB)
passes a null value which indicates that the cache should be
discarded. It will defer actually recalculating the cache until
necessary, e.g. due to a call to `navigator.storage.estimate()` or
another storage API querying the manager for overall usage.

On the benchmark mentioned in the bug, this change performs
equivalently to the "dumb" solution of completely ignoring quota
because the benchmark never triggers a usage recalculation.

This change also makes the simple but useful improvement of skipping
all quota update logic upon completion of a readonly transaction.

Bug: 1491201

Change-Id: Ic0354d3200af8cffc4870bc522a758d297a9ed7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4952750
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212455}