tast-tests/video: Add NV12 Scaling for GLImageProcessor
Previously the GLImageProcessor was not setup to do NV12 scaling on
Trogdor. Updates the GLImageProcessor to be able to perform NV12 scaling
by feeding in the planes as R8 and RG_88 buffers instead.

To verify this change, 5 new NV12 scaling tests were added. GLImageProcessor and LibYUV are tested by downscaling and upscaling NV12 frames and measuring the elapsed processing time. The GLImageProcessor's scaling accuracy is also tested by having it upscale and downscale an NV12 frame and verify and measure the scaled frame's PSNR compared to the original.

BUG: b:266785463
TEST:Ran ImageProcessorPerfTest on Trogdor and Asurada.

Change-Id: I6fb0ad0c9645f5df336faa7d218341b30a72caab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4769616
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212466}