Reland "[dPWA] Adding support for intent picker UX in WebAppIntegrationTestDriver (not on Lacros)"
This is a reland of commit 36f2c72f1cfbbf8298ed7f74c6293fd0090ac54f.
This does not contain a fix to the flakiness, but it does fix the underlying issue, i.e. the
WebAppIntegrationTestDriver does work properly for the new
intent picker UX (as is verified by running the tests failing in
crrev.com/c/4919875), just not on Lacros.

I could not find a reasoning behind why it was behaving this way, but
my findings have been documented in go/lacros-intent-ux-failure,
and can be tackled by adding support for lacros tests in
crbug.com/1462253.

Original change's description:
> [dPWA] Adding support for intent picker UX in WebAppIntegrationTestDriver
>
> This CL adds support for the new intent picker UX to be used in the
> WebAppIntegrationTestDriver by:
> 1. Piping whether the app icon will be shown all the way from the
> IntentPickerTabHelper to the test driver so that the latter can
> selectively wait if the icon needs to be shown in the intent chip.
> 2. Adds an observer that observes visibility for the intent chip
> button, and smartly determines if it needs to wait for the chip
> visibility to change.
>
> Verified by running existing WebAppIntegration tests that were
> failing locally once the flag was removed.
>
> Bug: 1412856, 1357905
> Change-Id: Ia376bccb2c2dcaaa5c4df35f9d91a3398f0830ba
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4942518
> Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
> Reviewed-by: Elly FJ <ellyjones@chromium.org>
> Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
> Reviewed-by: Glen Robertson <glenrob@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1211513}

Bug: 1412856, 1357905
Change-Id: I294910bc6f24c41ad2662ae0c749b7807ac27989
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4953565
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212497}