Reland "System Nudge: Add a close button to SystemNudgeView"
This is a reland of commit 0c6050e11ae182c4c93d9f540a9b371ae763f74d

There was a dangling ptr check failure due to assigning the results of
`widget->SetContentsView()` in the unit test to a raw pointer multiple
times. After re-assigning the results of `SetContentsView` all child
views are removed [1] which caused the dangling ptr.

The need to cache the contents view of the widget was removed by
accessing the required child view through its view ID directly.
A bug was filed (b/306466133) to do the same for other SystemNudgeView
child views, but a fix for this will be done in a separate CL.

[1] https://source.chromium.org/chromium/chromium/src/+/main:ui/views/widget/widget.cc;l=630;drc=430c7ea2e0a2e29142869f6efb112a7979ab605a

Original change's description:
> System Nudge: Add a close button to SystemNudgeView
>
> Add a close button that is visible on hover for nudges that are not
> text-only and do not have an anchor view.
>
> Whenever there is a close button, a new container view with a FillLayout
> is used to hold the `image_and_text_container` and the close button, to
> allow a bit of overlap which is required by the spec.
>
> specs: http://go/cros-educationalnudge-spec
> screenshot: https://screenshot.googleplex.com/5Ai64d4wMahttsH
>
> Bug: b:303684220
> Change-Id: I67a1f9301d45c6e5551cfb33f9933d77cc0d362e
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4948990
> Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
> Reviewed-by: James Cook <jamescook@chromium.org>
> Commit-Queue: Kevin Radtke <kradtke@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1211833}

Bug: b:303684220
Change-Id: I7537c7d26e389b3926994b29e76c0848cb3372f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4954824
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Commit-Queue: Kevin Radtke <kradtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212508}