Report full page process list in ReportPageProcessesPolicy
The page process list is sent to the resource manager daemon
(resourced). The page process list is used to estimate the memory usage
of the background and protected processes. Based on the Chrome memory
usage, resourced determines when to release memory from Chrome or VMs or
container. When there are a lot of background processes in Ash Chrome,
resourced avoids killing perceptible apps in VMs or container.

If the last page node event is within the miminum report interval, start
a timer to ensure the page process list is reported after the last page
node event.

Deprecate crosapi ReportBackgroundProcesses.

Bug: b/304679315
Change-Id: Id84c12ba5ea933470b5dffda3885083e98bb349e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4951554
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Vovo Yang <vovoy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212560}