Set app type to ash-chrome PiP
The app type of the ash-chrome PiP window isn't set on ChromeOS,
which causes it to be excluded from some WM logic on ChromeOS.

This is only problematic with ash-chrome PiP because for Lacros
and ARC PiP, Exo, which creates the window, sets up this property.
So, this property should also be set for ash-chrome PiP by the
component that initializes the window.

We should be able to remove this code once Lacros has fully
rolled out on ChromeOS.

Bug: b/306532421
Test: ash-chrome PiP isn't filtered out from ash window list.
Change-Id: I2cb9144c0e762695b05361e9470c9dfcc968cf29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4957279
Reviewed-by: Fr <beaufort.francois@gmail.com>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Commit-Queue: Kazuki Takise <takise@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212603}