Stop the blocking request for policy after the session is stopping
Sending a blocking D-Bus request after the session is stopping can
cause a dead-lock where session_manager is waiting for Chrome and
Chrome is waiting for response from D-Bus.

In this CL we subscribe to the signal to get when the session is
stopping and stop sending the blocking requests after that.

For now only requests to for the device policy is blocked
and if this will bring results, a follow up CL will stop sending
all other blocking requests.

BUG=1477697, 1475979, 836388
TEST=Manual test on device

Change-Id: Ie652d75bc2e43366b1b69aba0e82c42bb4897826
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4951739
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212631}