[iOS] Add context menu action histogram for Save to Photos
This CL adds a new IOS.SaveToPhotos.ContextMenu histogram to determine
what action is selected when the user long presses an image in a web
page when the "Save in Google Photos" action is presented, and when it
is not presented.

A new user action MobileWebContextMenuImageWithSaveToPhotosImpression is
added too so to count when the image context menu is impressed WITH the
"Save in Google Photos" action.

Low-Coverage-Reason: TRIVIAL_CHANGE logging statements
Bug: 1492924
Change-Id: I6f0720b563424d574c04f5ab4ec09c215ea84620
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4941832
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Cr-Commit-Position: refs/heads/main@{#1212633}