[ios] Move key LastInteractionTimeForGoodVisitsDiscover in PrefService
This CL moves LastInteractionTimeForGoodVisitsDiscover out of
NSUserDefaults.

The key has been moved in the PrefService. Note that there are two
types of PrefServices, in this case the code was moved in the
PrefService coming from BrowserState which is per account and is
syncable between devices.
Note that this kind of PrefService behaves differently from
NSUserDefaults which is per device (and not per account).

Bug: 1486770
Change-Id: I1f4cd08b30dbfd44284c2469f677f82545a61729
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4952532
Commit-Queue: Federica Germinario <fedegermi@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212714}