[Extensions] Add tests for the multiple extension uninstall dialog
Adds a unit test for the cancel callback in the multiple extensions
uninstall dialog. This test should verify that the error message is
properly shown and that the extensions are not removed. Adds a
browser/API test to verify that the uninstall dialog is properly shown
and that clicking the accept button removes the selected extensions.

Bug: 1432194
Change-Id: I33bd9d12a6643251348833cacb9c14c7e331e6b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4812457
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Zack Han <zackhan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212717}