Extract a function to prepare for load prefs for PrimaryAccountManager
This CL extracts to a separate method that prepares the preferences
before they are loaded by the PrimaryAccountManager. It also removes
all calls SetPrimaryAccountInternal when the prefs are being
pre-processed.

This CL should be a NO-OP compared to existing code and its reason
is just to make code more readable.

Bug: 1494295

Change-Id: I746c0968e908cadb7c4439ea875e805a829ac326
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4957138
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212778}