Extend PaintRenderingContext2D unit tests to validate recording
This will be useful in follow-up CLs to make sure the changes
there don't break anything.

Note that this recording validation reveals a bug in the unit test.
The test validating setTransform with a non-default scale factor
doesn't actually work, the test doesn't apply the scaling factor at
all. This will be fixed in a follow-up CL.

Bug: 1484741
Change-Id: Ib78d3308528f0c67f80bbd4606bc352a91d5e38a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4943652
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212780}