arc: Use GuestOsSessionTracker to get VmInfo for ARCVM
We currently store the VmInfo of the running ARCVM in ArcSessionManager
by making it receive VmStarted / VmStopped signals from vm_concierge,
but this is already done by GuestOsSessionTracker which is a
KeyedService. Hence this CL replaces the method
ArcSessionManager::GetVmInfo() with GuestOsSessionTracker::GetVmInfo()
and stops storing VmInfo in ArcSessionManager.

BUG=b:289330664
TEST=arc.Drivefs.vm, arc.ADBOverUSB.vm
TEST=unit_tests --gtest_filter="ArcSessionManagerTest.*"
TEST=unit_tests --gtest_filter="ArcAdbdMonitorBridgeTest.*"

Change-Id: I67fb3b273eed3fa32ffa41f041a2d317d8cc45b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4918991
Reviewed-by: Yury Khmel <khmel@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212798}