[ios] Hide feed and feed header depending on search engine choice
When the IOSHideFeedWithSearchChoice feature, the feed and feed header
should be hidden when/if a user selects a non-Google search engine as
their default.

This change also refactors the "isFeedHeaderVisible" method to move it
into the NTPMediator. This allows the mediator to own the business logic
and to make updates to the view as needed, which more closely matches
our design patterns. It also has the side effect of making it easier to
test this business logic!

Bug: 1493117
Change-Id: I72d425a8a29300652a5b09171f079d9568a962a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4950034
Commit-Queue: Scott Yoder <scottyoder@google.com>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1212809}