Suppress the privacy sandbox dialog during signed-in profile creation
https://crrev.com/c/4548474 tried to address the same problem but it
didn't take into account that the profile customization dialog is shown
asynchronously during the signed-in profile creation flow.

To fix the problem in that flow, we'd also like to suppress the
privacy sandbox dialog while the theme info is being downloaded from
Sync servers.

This a band-aid solution to the problem with multiple dialogs showing up
after the profile creation. We plan to address it with a more robust
solution that doesn't require dialogs to depend on each other in the
nearest future.

Bug: 1444751
Change-Id: I86e105dd0d6a84bbf7405c632fef86cdf2fd7f24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4955733
Reviewed-by: Nicola Tommasi <tommasin@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Auto-Submit: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Rainhard Findling <rainhard@chromium.org>
Commit-Queue: Nicola Tommasi <tommasin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212817}