Support aggregation coordinator selection for Private Aggregation
Adds a new aggregation_coordinator_origin field to the Private
Aggregation layer. This field is validated and then sent on to the
aggregation_service layer. This change is not web visible as the
integrations with Shared Storage and Protected Audience will be added in
separate cls.

See the related spec change:
https://github.com/patcg-individual-drafts/private-aggregation-api/pull/106

Note that this also modifies behavior of the database to only skip
reports that do not deserialize instead of skipping all reports if
there is any that do not deserialize. This may change behavior in case
of database corruption.

Bug: 1481254
Change-Id: Ie90eafc208f94ffd98ad95642b8f8c9b0af85fac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4885966
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Russ Hamilton <behamilton@google.com>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: Nan Lin <linnan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1212827}